Update index.html
Replaces links to frontend source code repositories with official links to repositories that are more friendly to the open source ecosystem.

Some explanations I feel may be needed:

The FreeTube PWA link has been changed to match the new repository name. This link change skips a redirect.

The Breezewiki link has been changed from the flagship instance to the Forgejo repo to follow the page's convention.

The above reasoning was not applied to Wikiless as the original repository now 404s with seemingly no alternatives from the original dev available.

The Simply Translate link has been changed from the SourceHut repo to the ManeraKai's Codeberg repo as development has changed hands.

The Send link was changed from GitHub to Gitlab, both because it's a more friendly option than GitHub, but also because the GitHub link was a mirror from Gitlab.

This fixes the issue of the link to Tubo 404'ing by replacing the now-defunct GitHub repo link with a link to the developer's cgit repo.